package/eudev: remove the prompt
The package eudev depends on DEVICE_CREATION_DYNAMIC_EUDEV, and
is also selected by DEVICE_CREATION_DYNAMIC_EUDEV.

This means that:
  - eudev is not visible when DEVICE_CREATION_DYNAMIC_EUDEV is not set
  - eudev is visible but forced when DEVICE_CREATION_DYNAMIC_EUDEV is set

As a consequence, eudev is never user-selectable.

Since the user already made the decision to use eudev in the /dev
management choice, there is no point in showing the eudev package at all
in the menuconfig.

Just remove the prompt.

Also remove the comments, since anyway the user can't do anything about it.

Although not strictly required, we anyway keep the dependencies expressed
in eudev, to explain where they come from (fork, util-linux...), so it is
explicit their duplication in the .dev management choice entry comes from
eudev itself. So, update the comments there.

Also, move the help text to the /dev macnagement choice, so the user has
a chance to look at it! ;-)

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Bernd Kuhls <bernd.kuhls@t-online.de>