package/systemd: remove prompt
The package systemd depends on BR2_INIT_SYSTEMD, and is also selected
by BR2_INIT_SYSTEMD.

This means that:
  - systemd is not visible when BR2_INIT_SYSTEMD is not set
  - systemd is visible but forced when BR2_INIT_SYSTEMD is set

As a consequence, systemd is never user-selectable.

Since the user already made the decision to use systemd in the init
system choice, there is no point in showing the systemd package at
all in the menuconfig.

Just remove the prompt, and move systemd's sub-option to a sub-menu.

Although not strictly required, we anyway keep the dependencies expressed
in systemd, to explain where they come from (kmod, util-linux...), so it
is explicit their duplication in the init system choice entry comes from
systemd itself. So, update the comments there.

Re-order dependencies, so they are the same in the systemd package and
the init system choice. Add dependencies missing in the package.

Also, move the help text to the init system choice, so the user has a
chance to look at it! ;-)

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Bernd Kuhls <bernd.kuhls@t-online.de>