infra: Improvement of the github helper
The github helper so far was only used to download generated tarball.
It could not be used for released tarball uploaded by the maintainer
which created some complexity on how to add a github package.

Now, by using the third parameter, either 'snapshot' or 'release' and
adjusting the fourth parameter, all packages can use the github helper.

Signed-off-by: "Yann E. Morin" <yann.morin.1998@free.fr>
Signed-off-by: Maxime Hadjinlian <maxime.hadjinlia@gmail.com>