openntpd: Fix missing home directory
We need to add a home directory for the _ntp users since ntpd will take the
home directory of its users as a directory it can chroot into.

Also, since '/var/run' points to '/tmp' in our skeleton, we need to
tell the init scripts to create the folder before starting ntpd.
This is also true with systemd, hence the uses of systemd-tmpfiles.

Signed-off-by: Maxime Hadjinlian <maxime.hadjinlian@gmail.com>