skeleton: New package
Create a proper package for the skeleton.

The main Makefile is modified to remove the skeleton support.
The 'dirs' target, will create the $(TARGET_DIR).
The file 'output/target/.root' doesn't exists anymore, as there's no
Make rule to statisfy.

The infrastructure are modified to filter host-skeleton.
It's needed becauses the host-dependencies are derived from the
dependencies of the target package where 'host-' is preprended to the
depedency name.

In the pkg-generic we add skeleton as a dependency to every package.

The whole system/system.mk is now removed at the profit of
package/skeleton/skeleton.mk

Signed-off-by: Maxime Hadjinlian <maxime.hadjinlian@gmail.com>
---
v4 -> v5:
    - Remove include system/system.mk from the main Makefile, the file
      doesn't exists anymore.
    - Filter out host-skeleton also from pkg-python and pkg-virtual
v3 -> v4:
    - Rework commit log
    - Modify if/else constructs for SKELETON_PATH
    - The 'dirs' target stays as it was
    - Empty system.mk into skeleton.mk, in the near future, we will want
      skeleton.mk to do everything on the default skeleton, even the
      getty stuff.
v2 -> v3:
   - Add patch from Thomas De Schampheleire so we create $(BUILD_DIR)
     before create $(BUILD_DIR)
v1 -> v2:
   - Huge overall rework
   - Filter-out host-skeleton in missing infra
   - Finally, keep skeleton's file in /system/
   - Keep the TARGET_DIR creation in the main Makefile