package: Introduce $(2)_INFRA_EXTRA_DEPENDENCIES
Instead of repeating the filter-out function to remove the host-package
that we don't want as dependencies in each infra, handle it it
pkg-generic through a generic variable, which is specified by each
infra.

The $(2)_INFRA_EXTRA_DEPENDENCIES contains the specific host package
that we want to filter from the package's dependencies.

Signed-off-by: Maxime Hadjinlian <maxime.hadjinlian@gmail.com>