zxing_cpp: New package
ZXing was hosted on Google Code, which is a closing forges.
The current link was already redirecting to the github repo.

But upstream is no longer supporting the cpp bindings, it's now a Java
library/application.

A fork has been created from the latests cpp source of the original
upstream.
So this is a new package from an old one and a legacy option has been
added.

Signed-off-by: Maxime Hadjinlian <maxime.hadjinlian@gmail.com>